backend/requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.4
Django==4.1.1
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.13.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.1.0
pytz==2022.2.1
sqlparse==0.4.2
Werkzeug==2.2.2
